backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
ecdsa==0.19.1
emoji==2.14.1
faiss-cpu==1.10.0
fastapi==0.115.12
filetype==1.2.0
Flask==3.1.1
fonttools==4.57.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.62
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.24
limits==4.6
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
multidict==6.3.2
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.4
olefile==0.47
openai==1.70.0
orjson==3.10.16
packaging==24.2
passlib==1.7.4
pillow==11.1.0
propcache==0.3.1
psutil==7.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
pypdf==5.5.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.11.2
python-iso639==2025.2.18
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-socketio==5.12.1
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
unstructured==0.17.2
unstructured-client==0.36.0
urllib3==2.3.0
uv==0.7.2
uvicorn==0.34.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
